requirements.txt
alembic==1.15.2
aniso8601==10.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
chardet==5.2.0
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.2.1
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
reportlab==4.3.1
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
